commit jwt

+++ controllers/api/v1/posts_api.js
@@ -15,10 +15,18 @@ module.exports.index=async function(req,res)
 module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findByIdAndDelete(req.params.id);
+        if(post.user==req.user.id){  //after generating the token user will automaticallly add in req object  this is passport feature
             await Comment.deleteMany({post: req.params.id});
              return res.json(200,{
                 message:"Post and associated comments deleted"
              })
+            }
+            else
+            {
+                return res.json(401,{
+                    message:'You cannot delete this post'   
+                }) 
+            }
     }
     catch(err){
